+++ diandian/src/main/java/com/aizhixin/cloud/dd/questionnaire/dto/QuestionDTO.java
@@ -29,7 +29,7 @@ public class QuestionDTO {
      * 题目分值
      */
     @ApiModelProperty(value = "题目分值", required = false)
-    private Integer score = 0;
+    private Float score = 0f;
 
     /**
      * 选择题选项
@@ -41,7 +41,7 @@ public class QuestionDTO {
      * 题目分值
      */
     @ApiModelProperty(value = "实际得分", required = false)
-    private Integer actualScore = 0;
+    private Float actualScore = 0f;
 
     /**
      * 学生答案
@@ -58,7 +58,7 @@ public class QuestionDTO {
     public QuestionDTO() {
     }
 
-    public QuestionDTO(Long id, Integer no, String name, Integer score, Integer actualScore) {
+    public QuestionDTO(Long id, Integer no, String name, Float score, Float actualScore) {
         super();
         this.id = id;
         this.no = no;
